+++ messages/javascript/src/index.ts
@@ -1,8 +1,4 @@
 import * as $protobuf from 'protobufjs'
-import MessageToBinaryStream from './MessageToBinaryStream'
-import MessageToNdjsonStream from './MessageToNdjsonStream'
-import BinaryToMessageStream from './BinaryToMessageStream'
-import NdjsonToMessageStream from './NdjsonToMessageStream'
 import * as TimeConversion from './TimeConversion'
 import * as IdGenerator from './IdGenerator'
 import { io } from './messages'
@@ -12,13 +8,4 @@ import { version } from '../package.json'
 $protobuf.util.Long = undefined
 $protobuf.configure()
 
-export {
-  messages,
-  MessageToBinaryStream,
-  MessageToNdjsonStream,
-  BinaryToMessageStream,
-  NdjsonToMessageStream,
-  TimeConversion,
-  IdGenerator,
-  version,
-}
+export { messages, TimeConversion, IdGenerator, version }